Add files via upload
 Lateral spine ova of Schistosoma in urine of subjects in Abuja Nigeria
Details: Schistosomiasis is a snail-borne acute and chronic parasitic disease that is caused by trematode blood flukes of the genus Schistosoma. There are two form in human (intestinal and urinary form). Laboratory diagnosis of the intestinal form of the disease is by microscopic identification of a lateral spined ova with well-defined shell outline and space between shell in the feaces of its host while the diagnosis of the urogenital form is by identification of a conspicuous terminal spine ova that contain a mature miracidium in the urine or biopsy of its host. Occurrence of any form of these ova in otherwise specimen were believed to have been as a result of contamination. We observed an unusual presence of lateral spine ova of Schistosoma in the urine of subjects within the Federal Capital Territory, North Central Nigeria. We followed up to ensure it was not contamination by allowing the male subject to produce a clean catch urine under supervision. The molecular analysis produced genes that align with S.heamatobium and other species known to infect animals such as cattles and sheeps. Although ova had a morphological resemblance with S. mansoni, gene alignment did not match this species